Make it compile on all supported Go platforms
This was done by factoring platform specific functions and constants
into their own files with build constraints.

This also makes sure that if FileInfo.Sys() returns nil it doesn't
crash.

The compileall.sh shell script can be used to test that it compiles on
all platforms.  This has been added to the travis tests.